requirements.txt
Django==1.7
django-dashing==0.2.6
widget-party==2.1
